# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZENOVA Platform API",
    "description": "Research-Grade Mental-Health Wellbeing and Conversational Support Platform API",
    "contact": {
      "name": "ZENOVA Engineering & Clinical Team",
      "url": "https://zenova.ai/",
      "email": "support@zenova.ai"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/fusion/fuse": {
      "post": {
        "tags": [
          "Multimodal Fusion"
        ],
        "summary": "Fuse Modalities",
        "description": "Fuse heterogeneous multimodal inputs into a unified clinical affective assessment.",
        "operationId": "fuse_modalities_api_v1_fusion_fuse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FusionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FusedMultimodalState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fusion/evaluate": {
      "get": {
        "tags": [
          "Multimodal Fusion"
        ],
        "summary": "Evaluate Multimodal Fusion",
        "description": "Execute the comparative evaluation benchmark comparing Text-Only, Rule Fusion, and Learnable GMU.",
        "operationId": "evaluate_multimodal_fusion_api_v1_fusion_evaluate_get",
        "parameters": [
          {
            "name": "num_samples",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 5,
              "description": "Evaluation samples per modality regime",
              "default": 30,
              "title": "Num Samples"
            },
            "description": "Evaluation samples per modality regime"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparativeEvaluationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orchestrator/process": {
      "post": {
        "tags": [
          "Orchestrator"
        ],
        "summary": "Process Turn",
        "description": "High-level unified API for end-to-end turn processing across all stages.",
        "operationId": "process_turn_api_v1_orchestrator_process_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrchestratorTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrchestrationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/traces/{trace_id}": {
      "get": {
        "tags": [
          "Orchestrator"
        ],
        "summary": "Get Execution Trace",
        "description": "Inspect full internal execution trace for authorized clinicians or developers.",
        "operationId": "get_execution_trace_api_v1_orchestrator_traces__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician_admin",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineTrace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orchestrator/traces": {
      "get": {
        "tags": [
          "Orchestrator"
        ],
        "summary": "List Execution Traces",
        "description": "List execution traces filtered by session, user, or status for authorized clinicians.",
        "operationId": "list_execution_traces_api_v1_orchestrator_traces_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician_admin",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparativeEvaluationReport": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "total_eval_samples": {
            "type": "integer",
            "title": "Total Eval Samples"
          },
          "models": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModelComparisonMetrics"
            },
            "type": "object",
            "title": "Models"
          },
          "summary_findings": {
            "type": "string",
            "title": "Summary Findings"
          },
          "multimodal_improves_performance": {
            "type": "boolean",
            "title": "Multimodal Improves Performance"
          },
          "best_overall_model": {
            "type": "string",
            "title": "Best Overall Model"
          }
        },
        "type": "object",
        "required": [
          "total_eval_samples",
          "models",
          "summary_findings",
          "multimodal_improves_performance",
          "best_overall_model"
        ],
        "title": "ComparativeEvaluationReport",
        "description": "Comparative evaluation benchmark report comparing Text-Only, Rule Fusion, and Learnable GMU."
      },
...
      "EmotionCategory": {
        "type": "string",
        "enum": [
          "joy",
          "sadness",
          "anger",
          "anxiety",
          "fear",
          "neutral",
          "frustration",
          "hope",
          "grief",
          "guilt",
          "shame"
        ],
        "title": "EmotionCategory"
      },
...
      "FusedMultimodalState": {
        "properties": {
          "primary_affect": {
            "$ref": "#/components/schemas/EmotionCategory"
          },
          "fused_valence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Fused Valence",
            "description": "Calibrated valence (-1.0 negative to +1.0 positive)"
          },
          "fused_arousal": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fused Arousal",
            "description": "Calibrated arousal (0.0 calm to 1.0 excited/agitated)"
          },
          "fused_dominance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Fused Dominance",
            "description": "Calibrated dominance",
            "default": 0.0
          },
          "fused_distress_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fused Distress Score",
            "description": "Unified distress severity index (0.0 to 1.0)"
          },
          "fused_risk_level": {
            "$ref": "#/components/schemas/RiskLevel",
            "description": "Unified crisis and triage risk tier"
          },
          "urgency_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Urgency Score",
            "description": "Overall intervention urgency (0.0 to 1.0)"
          },
          "modality_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Modality Weights",
            "description": "Normalized contribution weights of each modality"
          },
          "modality_mask": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Modality Mask",
            "description": "Availability mask at time of fusion"
          },
          "discrepancy": {
            "$ref": "#/components/schemas/CrossModalDiscrepancy"
          },
          "fusion_method": {
            "type": "string",
            "title": "Fusion Method",
            "description": "'weighted_rule' or 'learnable_gmu'"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Overall confidence of the fused assessment"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "primary_affect",
          "fused_valence",
          "fused_arousal",
          "fused_distress_score",
          "fused_risk_level",
          "urgency_score",
          "modality_weights",
          "modality_mask",
          "fusion_method",
          "confidence"
        ],
        "title": "FusedMultimodalState",
        "description": "Unified multimodal affective, clinical, and risk assessment."
      },
      "FusionRequest": {
        "properties": {
          "user_input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "emotion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmotionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "symptoms": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SymptomResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "voice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoiceResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "behavior": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BehavioralResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "baseline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BaselineResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider",
            "description": "'weighted_rule' or 'learnable_gmu'",
            "default": "weighted_rule"
          }
        },
        "type": "object",
        "title": "FusionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrchestrationResult": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "turn_id": {
            "type": "integer",
            "title": "Turn Id"
          },
          "user_input": {
            "type": "string",
            "title": "User Input"
          },
          "response": {
            "type": "string",
            "title": "Response"
          },
          "escalated_to_human": {
            "type": "boolean",
            "title": "Escalated To Human",
            "default": false
          },
          "escalation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Id"
          },
          "escalation_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Event Id"
          },
          "emotion": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion"
          },
          "symptoms": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symptoms"
          },
          "risk": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk"
          },
          "baseline": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline"
          },
          "behavior": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behavior"
          },
          "voice": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          },
          "fused_state": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fused State"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "strategy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "rag": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rag"
          },
          "safety": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "pipeline_status": {
            "$ref": "#/components/schemas/PipelineStatus",
            "default": "nominal"
          },
          "degraded_modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degraded Modules"
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "db_persisted": {
            "type": "boolean",
            "title": "Db Persisted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "turn_id",
          "user_input",
          "response"
        ],
        "title": "OrchestrationResult",
        "description": "Comprehensive typed result from end-to-end turn processing."
      },
...
      "OrchestratorTurnRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "audio_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Base64"
          },
          "passive_behavior_metrics": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passive Behavior Metrics"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id"
        ],
        "title": "OrchestratorTurnRequest",
        "description": "Request payload for the high-level orchestrator process API."
      },
...
      "PipelineTrace": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "turn_id": {
            "type": "integer",
            "title": "Turn Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "status": {
            "$ref": "#/components/schemas/PipelineStatus",
            "default": "nominal"
          },
          "total_duration_ms": {
            "type": "number",
            "title": "Total Duration Ms",
            "default": 0.0
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/TraceSpan"
            },
            "type": "array",
            "title": "Spans"
          },
          "degraded_modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degraded Modules"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "session_id",
          "turn_id",
          "user_id"
        ],
        "title": "PipelineTrace",
        "description": "Immutable execution trace of a conversational turn across all pipeline stages."
      },
...
    }
  },
...
}